pyproject.toml
[project]
name = "agno5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq>=0.24.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.0",
]
